requirements.txt
astroid==2.4.1
boto3==1.12.46
botocore==1.15.46
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
docutils==0.15.2
git+ssh://git@github.com/eunimart/vdezi_event_handler.git
idna==2.9
isort==4.3.21
jmespath==0.9.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
mongoengine==0.19.1
msgpack-python==0.5.6
pika==1.1.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.2
pymongo==3.10.1
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2019.3
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
toml==0.10.1
typed-ast==1.4.1
ua-parser==0.10.0
umsgpack==0.1.0
urllib3==1.25.9
user-agents==2.1
watchtower==0.7.3
wrapt==1.12.1
xmltodict==0.12.0
